fix(actions): le repos échouait sur « Unknown column 'r' »
L'action repos portait deux instructions pour le malus : celle de type
rest, qui calcule la récupération, et une instruction player dont la
valeur était la lettre « r » — un nom de trait là où un entier est
attendu. Player::putBonus collait la valeur telle quelle dans son
INSERT, et MariaDB la lisait comme un nom de colonne.

putBonus lie désormais ses valeurs en paramètres préparés, entières ;
PlayerOutcomeInstruction convertit sa valeur en entier comme son
schéma l'annonce. La migration retire l'instruction fautive par son
contenu, elle n'avait rien à ajouter ; le dump est aligné.

+++ src/Action/OutcomeInstruction/PlayerOutcomeInstruction.php
@@ -31,7 +31,9 @@ class PlayerOutcomeInstruction extends OutcomeInstruction implements HasParamete
         
         $player = $params['player'] ?? null;
         $carac = $params['carac'] ?? null;
-        $value = $params['value'] ?? 0;
+        // The schema declares an INT; the admin form still stores what it is
+        // given, so a stray string becomes 0 rather than a broken query.
+        $value = (int) ($params['value'] ?? 0);
         $outcomeSuccessMessages = array();
 
         if ($carac !== null && in_array($player, ['actor', 'target'], true)) {
